requirements.txt
python-telegram-bot
PyPDF2
pdf2image
Pillow
pytesseract
ffmpeg-python
flask
gunicorn